Back-end/ms_compras/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.2
marshmallow==3.19.0
google-cloud-pubsub==2.13.11
psycopg2-binary==2.9.6
pytest==7.2.2